+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## next
 
 - Refined OpenAPI doc generation to output only non-null attributes in the InfoObject.
+- Fixed filtering params validation to properly allow null values for the "!" and "!!" operators
 
 ## 2.0.pre.6